+++ application/common/config.php
@@ -1,4 +1,4 @@
 <?php
-return [ 
-		'version' => "V20190829.common",
+return [
+	'version' => "V20190829.common",
 ];